+++ Assets/Scripts/World.cs
@@ -75,23 +75,23 @@ public class World : MonoBehaviour
         int cx, cy, cz;
 
         if (pos.x < 0)
-            cx = (int)((Mathf.Round(pos.x - chunkSize) + 1) / (float)chunkSize) * chunkSize;
+            cx = (int)((Mathf.Round(pos.x - chunkSize) + 1) / (double)chunkSize) * chunkSize;
         else
-            cx = (int)(Mathf.Round(pos.x) / (float)chunkSize) * chunkSize;
+            cx = (int)(Mathf.Round(pos.x) / (double)chunkSize) * chunkSize;
 
         if (pos.y < 0)
-            cy = (int)((Mathf.Round(pos.y - chunkSize) + 1) / (float)chunkSize) * chunkSize;
+            cy = (int)((Mathf.Round(pos.y - chunkSize) + 1) / (double)chunkSize) * chunkSize;
         else
-            cy = (int)(Mathf.Round(pos.y) / (float)chunkSize) * chunkSize;
+            cy = (int)(Mathf.Round(pos.y) / (double)chunkSize) * chunkSize;
 
         if (pos.z < 0)
-            cz = (int)((Mathf.Round(pos.z - chunkSize) + 1) / (float)chunkSize) * chunkSize;
+            cz = (int)((Mathf.Round(pos.z - chunkSize) + 1) / (double)chunkSize) * chunkSize;
         else
-            cz = (int)(Mathf.Round(pos.z) / (float)chunkSize) * chunkSize;
+            cz = (int)(Mathf.Round(pos.z) / (double)chunkSize) * chunkSize;
 
-        int blx = (int)Mathf.Abs((float)Mathf.Round(pos.x) - cx);
-        int bly = (int)Mathf.Abs((float)Mathf.Round(pos.y) - cy);
-        int blz = (int)Mathf.Abs((float)Mathf.Round(pos.z) - cz);
+        int blx = (int)Mathf.Abs((int)Mathf.Round(pos.x) - cx);
+        int bly = (int)Mathf.Abs((int)Mathf.Round(pos.y) - cy);
+        int blz = (int)Mathf.Abs((int)Mathf.Round(pos.z) - cz);
 
         string cn = BuildChunkName(new Vector3(cx, cy, cz));
         Chunk c;
@@ -266,7 +266,7 @@ public class World : MonoBehaviour
             {
                 if (closedSet.Contains(neighbourBlock))
                     continue;
-                float newMovementCostToNeighbour = currentBlock.gCost + calcDistance(currentBlock, neighbourBlock);
+                double newMovementCostToNeighbour = (double)(currentBlock.gCost + calcDistance(currentBlock, neighbourBlock));
                 if (newMovementCostToNeighbour < neighbourBlock.gCost || !openSet.Contains(neighbourBlock))
                 {
                     neighbourBlock.gCost = newMovementCostToNeighbour;
@@ -333,10 +333,10 @@ public class World : MonoBehaviour
     /// <summary>
     /// calculates the distance between two blocks along the x- and z-axis, where only block-to-direct-neighbourblock-movement is allowed
     /// </summary>
-    public float calcDistance(Block blockA, Block blockB)
+    public double calcDistance(Block blockA, Block blockB)
     {
-        float dstX = Mathf.Abs(blockA.worldPosition.x - blockB.worldPosition.x);
-        float dstY = Mathf.Abs(blockA.worldPosition.z - blockB.worldPosition.z);
+        double dstX = Mathf.Abs(blockA.worldPosition.x - blockB.worldPosition.x);
+        double dstY = Mathf.Abs(blockA.worldPosition.z - blockB.worldPosition.z);
 
         //moving diagonally to a neighbourblock results in a distance of roughly 1.41 worldunits
         if (dstX > dstY)
@@ -347,10 +347,11 @@ public class World : MonoBehaviour
 
 
     /// <summary>
-    /// retraces the path from end to start by going along the pathParents of each block
+    /// 
+    /// s the path from end to start by going along the pathParents of each block
     /// </summary>
     /// <param name="cost"> the total cost of the path </param>
-    public Blockpath RetracePath(Block startBlock, Block endBlock, float cost, Heap<Block> openSet, HashSet<Block> closedSet)
+    public Blockpath RetracePath(Block startBlock, Block endBlock, double cost, Heap<Block> openSet, HashSet<Block> closedSet)
     {
         List<Block> blockList = new List<Block>();
         Block currentBlock = endBlock;
@@ -371,8 +372,8 @@ public class World : MonoBehaviour
         }
         foreach (Block block in closedSet)
             block.ResetPathfindingVal();
-
-        Blockpath result = new Blockpath(startBlock, endBlock, cost, blockList);
+        
+        Blockpath result = new Blockpath(startBlock, endBlock, (float)cost, blockList);
 
         if (startBlock.edges == null)
             startBlock.edges = new List<Blockpath>();
@@ -536,7 +537,7 @@ public class World : MonoBehaviour
     }*/
 
 
-    float timer = 2f;
+    double timer = 2f;
     public GameObject testCube;
 
     bool einmal = true;
@@ -666,8 +667,8 @@ public class World : MonoBehaviour
                 UnityEngine.Debug.Log("random apple spawn failed to spawn all apples");
                 return;
             }
-            float xOffset = Random.Range(5f, 60f);
-            float zOffset = Random.Range(5f, 60f);
+            double xOffset = Random.Range(5f, 60f);
+            double zOffset = Random.Range(5f, 60f);
 
             Vector3 spawnPos = new Vector3(10f + (int)xOffset, 65f, 10f + (int)zOffset);
